add sample_stage DB

+++ scripts/otomo/analysis_status.py
@@ -5,7 +5,7 @@ import glob
 import os
 import otomo.CONFIG
 
-def get_sample_w_status(status, where = None, limit = 0, conf_file=otomo.CONFIG.DEFAULT_CONF):
+def get_sample_w_status(status, limit = 0, conf_file=otomo.CONFIG.DEFAULT_CONF):
     conf = otomo.CONFIG.load_conf(conf_file)
     db = conf.get("db", "analysis_db")
 
@@ -13,21 +13,9 @@ def get_sample_w_status(status, where = None, limit = 0, conf_file=otomo.CONFIG.
     if limit > 0:
         sql_limit = "limit %d" % (limit)
 
-    data = {"status": status}
-    sql_where = ""
-    if where != None:
-        """
-        where = {
-            "sql": "description=:desc and run_count>:count",
-            "value": {"desc": "join", "count": 2}
-        }
-        """
-        sql_where = "and (%s)" % (where["sql"])
-        data.update(where["value"])
-
     con = sqlite3.connect(db)
     cur = con.cursor()
-    cur.execute("select sample from analysis where status=:status %s %s" % (sql_where, sql_limit), data)
+    cur.execute("select sample from analysis where status=:status %s" % (sql_limit), {"status": status})
     
     samples = []
     for f in cur.fetchall():
@@ -36,72 +24,66 @@ def get_sample_w_status(status, where = None, limit = 0, conf_file=otomo.CONFIG.
     con.close()
     return samples
 
-def get_sample_count_g_status(where = None, conf_file=otomo.CONFIG.DEFAULT_CONF):
+def get_sample_count_groupby_status_stage(conf_file=otomo.CONFIG.DEFAULT_CONF):
     conf = otomo.CONFIG.load_conf(conf_file)
     db = conf.get("db", "analysis_db")
-
-    data = {}
-    sql_where = ""
-    if where != None:
-        """
-        where = {
-            "sql": "description=:desc and run_count>:count",
-            "value": {"desc": "join", "count": 2}
-        }
-        """
-        sql_where = "where " + where["sql"]
-        data.update(where["value"])
-
     con = sqlite3.connect(db)
     cur = con.cursor()
-    cur.execute("select status,count(sample) from analysis %s group by status" % (sql_where), data)
+    cur.execute("select status,stage,count(sample) from analysis group by status,stage")
     
-    status = {}
-    for f in cur.fetchall():
-        status[f[0]] = f[1]
-
+    ret = {}
+    for status,stage,count in cur.fetchall():
+        if not status in ret:
+            ret[status] = {}
+        if not stage in ret[status]:
+            ret[status][stage] = count
+    
     con.close()
-    return status
+    return ret
 
 def get_run_count_w_sample(sample, conf_file=otomo.CONFIG.DEFAULT_CONF):
     conf = otomo.CONFIG.load_conf(conf_file)
-    db = conf.get("db", "analysis_db")
+    db = conf.get("db", "sample_stage_db")
 
     con = sqlite3.connect(db)
     cur = con.cursor()
-    cur.execute("select run_count from analysis where sample=:sample", {"sample": sample})
+    cur.execute("select stage,count(*) from sample_stage where sample=:sample group by stage", {"sample": sample})
 
-    count = 0
+    count = {}
     for f in cur.fetchall():
-        count = f[0]
-
+        count[f[0]] = f[1]
+    
     con.close()
     return count
 
-def set_status_request(sample, status, countup = False, description = "", error_text = "", stop_reason = "", conf_file=otomo.CONFIG.DEFAULT_CONF):
+def set_status_request(sample, status, stage = "", error_text = "", stop_reason = "", conf_file=otomo.CONFIG.DEFAULT_CONF):
 
     conf = otomo.CONFIG.load_conf(conf_file)
-    db = conf.get("db", "analysis_db")
-
+    
     now = datetime.datetime.now()
+    label_time = now.strftime('%Y%m%d_%H%M%S%f')
+    db_time = otomo.CONFIG.date_to_text(now)
+
     write_data = {
         "sample": sample,
         "status": status,
-        "last_update": otomo.CONFIG.date_to_text(now)
+        "last_update": db_time
     }
-    if countup:
-        con = sqlite3.connect(db)
-        cur = con.cursor()
-        cur.execute("select run_count from analysis where sample='%s'" % (sample))
 
-        run_count = 0
-        for f in cur.fetchall():
-            run_count = f[0]
-        con.close()
-        write_data["run_count"] = run_count + 1
+    def set_sample_job_request():
+        write_data = {
+            "sample": sample,
+            "stage": stage,
+            "start_time": db_time
+        }
+        temp_name = "%s/sample_stage_%s_%s.json" % (conf.get("db", "request_dir"), label_time, sample)
+        fw = open(temp_name, "w")
+        json.dump(write_data, fw)
+        fw.close()
 
-    if description != "":
-        write_data["description"] = description
+    if stage != "":
+        write_data["stage"] = stage
+        set_sample_job_request()
 
     if error_text != "":
         write_data["error_text"] = error_text
@@ -109,16 +91,17 @@ def set_status_request(sample, status, countup = False, description = "", error_
     if stop_reason != "":
         write_data["stop_reason"] = stop_reason
 
-    temp_name = "%s/%s_%s.json" % (conf.get("db", "request_dir"), now.strftime('%Y%m%d_%H%M%S%f'), sample)
+    temp_name = "%s/analysis_%s_%s.json" % (conf.get("db", "request_dir"), label_time, sample)
     fw = open(temp_name, "w")
     json.dump(write_data, fw)
     fw.close()
 
 def set_status_commit(conf_file=otomo.CONFIG.DEFAULT_CONF):
 
+    # analysis db
     conf = otomo.CONFIG.load_conf(conf_file)
     db = conf.get("db", "analysis_db")
-    requests = sorted(glob.glob("%s/*.json" % (conf.get("db", "request_dir"))))
+    requests = sorted(glob.glob("%s/analysis_*.json" % (conf.get("db", "request_dir"))))
     sqls = []
     commited_requests = []
     for i,request in enumerate(requests):
@@ -135,35 +118,63 @@ def set_status_commit(conf_file=otomo.CONFIG.DEFAULT_CONF):
         sqls.append(("update analysis set %s where sample=:sample" % (",".join(items)), data))
         commited_requests.append(request)
 
-    error = None
-    try:
-        con = sqlite3.connect(db, isolation_level='EXCLUSIVE')
-        cur = con.cursor()
-        for sql in sqls:
-            #print(sql)
-            cur.execute(sql[0], sql[1])
-        con.commit()
+    if len(sqls) > 0:
+        error = None
+        try:
+            con = sqlite3.connect(db, isolation_level='EXCLUSIVE')
+            cur = con.cursor()
+            for sql in sqls:
+                cur.execute(sql[0], sql[1])
+            con.commit()
+        except Exception as e:
+            error = e
+        finally:
+            if con:
+                con.close()
+        if error != None:
+            raise error
 
-    except Exception as e:
-        error = e
+        for request in commited_requests:
+            os.remove(request)
 
-    finally:
-        if con:
-            con.close()
+    # sample_stage db
+    db = conf.get("db", "sample_stage_db")
+    requests = sorted(glob.glob("%s/sample_stage_*.json" % (conf.get("db", "request_dir"))))
+    insert_list = []
+    commited_requests = []
+    for i,request in enumerate(requests):
+        if i >= 1000:
+            break
+        f = open(request)
+        data = json.load(f)
+        f.close()
+        insert_list.append((data["sample"], data["stage"], data["start_time"]))
+        commited_requests.append(request)
 
-    if error != None:
-        raise error
+    if len(insert_list) > 0:
+        error = None
+        try:
+            con = sqlite3.connect(db, isolation_level='EXCLUSIVE')
+            cur = con.cursor()
+            cur.executemany("insert into sample_stage values (?, ?, ?)", insert_list)
+            con.commit()
+        except Exception as e:
+            error = e
+        finally:
+            if con:
+                con.close()
+        if error != None:
+            raise error
 
-    for request in commited_requests:
-        os.remove(request)
+        for request in commited_requests:
+            os.remove(request)
 
 def main(args):
     if args.commit == False:
         set_status_request(
             args.sample, 
             args.status, 
-            countup = args.countup, 
-            description = args.description, 
+            stage = args.stage, 
             error_text = args.error_text, 
             stop_reason = args.stop_reason, 
             conf_file=args.conf
